validation

+++ read.php
@@ -1,10 +1,6 @@
 <?php
 session_start();
 require __DIR__ . '../functions.php';
-if ($_SESSION['login'] !== 'logIn') {
-    header('Location: http://localhost/bank_php/auth/login.php');
-    die;
-}
 ?>
 
 <!DOCTYPE html>
@@ -15,19 +11,19 @@ if ($_SESSION['login'] !== 'logIn') {
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/css/bootstrap.min.css">
     <script src="http://localhost/bank_php/script.js" defer></script>
-    
+
     <title>Read</title>
 </head>
 
 <body>
-    <?php require __DIR__ . '/menu.php'?>
+    <?php require __DIR__ . '/menu.php' ?>
     <?php require __DIR__ . '/msg.php' ?>
 
 
     <div class="container mt-5">
         <div class="row">
             <div class="col">
-                <h2>Account List</h2>
+                <h2>Accounts</h2>
             </div>
         </div>
     </div>
@@ -50,22 +46,26 @@ if ($_SESSION['login'] !== 'logIn') {
                         <b>Balance</b>
                     </div>
 
-                    <div class="col-3">
-                        <b>Action</b>
-                    </div>
+                    
+                    <?php if (isset($_SESSION['login']) && $_SESSION['login'] == 'logIn') : ?>
+                        <div class="col-3">
+                            <b>Action</b>
+                        </div>
+                    <?php endif ?>
+
                 </div>
             </div>
         </li>
-        <?php $members = unserialize(file_get_contents(__DIR__.'/data/users.ser')) ?>
+        <?php $members = unserialize(file_get_contents(__DIR__ . '/data/users.ser')) ?>
 
         <?php usort($members, fn ($a, $b) => $a['lastname'] <=> $b['lastname']) ?>;
-      
+
         <?php foreach ($members as $member) : ?>
             <li class="list-group-item">
                 <div class="container">
                     <div class="row">
                         <div class="col-2">
-                            <?= $member['name']. " ".$member['lastname'] ?>
+                            <?= $member['name'] . " " . $member['lastname'] ?>
                         </div>
                         <div class="col-2">
                             <?= $member['personalCode'] ?>
@@ -76,11 +76,15 @@ if ($_SESSION['login'] !== 'logIn') {
                         <div class="col-2">
                             <?= $member['balance'] ?>
                         </div>
-                        <div class="col-3">
-                            <a href="http://localhost/bank_php/addMoney.php?id=<?= $member['id'] ?>" class="btn btn-outline-success btn-sm">Add</a>
-                            <a href="http://localhost/bank_php/withdraw.php?id=<?= $member['id'] ?>" class="btn btn-outline-info btn-sm">Debit</a>
-                            <a href="http://localhost/bank_php/delete.php?id=<?= $member['id'] ?>" class="btn btn-outline-danger btn-sm">Delete</a>
-                        </div>
+
+                        <?php if (isset($_SESSION['login']) && $_SESSION['login'] == 'logIn') : ?>
+                            <div class="col-3">
+                                <a href="http://localhost/bank_php/addMoney.php?id=<?= $member['id'] ?>" class="btn btn-outline-success btn-sm">Add</a>
+                                <a href="http://localhost/bank_php/withdraw.php?id=<?= $member['id'] ?>" class="btn btn-outline-info btn-sm">Debit</a>
+                                <a href="http://localhost/bank_php/delete.php?id=<?= $member['id'] ?>" class="btn btn-outline-danger btn-sm">Delete</a>
+                            </div>
+                        <?php endif ?>
+
                     </div>
                 </div>
             </li>
